Update README.md
Aquaman is a 2018 American superhero film based on the DC Comics character of the same name. Directed by James Wan, it stars Jason Momoa as the titular character, with Amber Heard, Willem Dafoe, Patrick Wilson, Dolph Lundgren, Yahya Abdul-Mateen II and Nicole Kidman in supporting roles. It tells the story of Arthur Curry's journey to discover his identity as the heir to the underwater kingdom of Atlantis and his attempt to save both worlds from a war between surface dwellers and Atlanteans. With its stunning visuals and intense action sequences, Aquaman is sure to be one of the most talked about movies of 2018.